# bpo-38045: Improve the performance of _decompose() in the enum.py