--- a/Misc/NEWS.d/next/Library/2019-09-30-12-09-41.bpo-38045.VDRtd3.rst
+++ b/Misc/NEWS.d/next/Library/2019-09-30-12-09-41.bpo-38045.VDRtd3.rst
@@ -1,1 +1,1 @@
-Improve the performance of _decompose() in the enum.py. Patch by hongweipeng.
+Improve the performance of :func:`enum._decompose` in :mod:`enum`. Patch by hongweipeng.