The original code will access each element of `_value2member_map_` and check if the value is the power of 2.

After optimization, access the elements in `_value2member_map_` as needed. In the initial example:
```
from enum import Flag
F = Flag("F", list("abcdefghijklm"))
for idx in range(2**len(F) - 1):
    F(idx)
```
the speedup is more than 100x .(The same for replacing Flag with IntFlag).

<!-- issue-number: [bpo-38045](https://bugs.python.org/issue38045) -->
https://bugs.python.org/issue38045
<!-- /issue-number -->