@corona10 2019-09-30
@hongweipeng  Can you show me the benchmark result?