@hongweipeng 2019-09-30
Use the initial example:
```
from enum import Flag
F = Flag("F", list("abcdefghijklm"))
for idx in range(2**len(F) - 1):
    F(idx)
```
Before patch:
```
# time ./python ts.py

real	0m29.639s
user	0m29.615s
sys	0m0.020s
```
After patch:
```
# time ./python ts.py

real	0m0.135s
user	0m0.122s
sys	0m0.014s
```