@corona10 2019-09-30
@ethanfurman 
 
For core-developers,
I didn't check this PR is the correct way.
But seems like there is performance enhancement with this PR.

I left the benchmark for the core-developers

## Baseline
```` bash
 ./python.exe -m pyperf timeit -s 'from enum import Flag' 'F = Flag("F", list("abcdefghijklm"))' 'for idx in range(100):' '   F(idx)'
.....................
Mean +- std dev: 24.5 ms +- 0.3 ms
````

## PR
````
./python.exe -m pyperf timeit -s 'from enum import Flag' 'F = Flag("F", list("abcdefghijklm"))' 'for idx in range(100):' '   F(idx)'
.....................
Mean +- std dev: 7.30 ms +- 0.19 ms
````